Feature/create chromosome

+++ src/main/java/com/breaze/genomebank/function/infraestructure/repository/IFunctionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IFunctionRepository extends JpaRepository<Function,Long> {
+public interface IFunctionRepository extends JpaRepository<Function,Integer> {
 }
